+++ hcatalog/core/src/main/java/org/apache/hive/hcatalog/common/HiveClientCache.java
@@ -47,11 +47,11 @@ import org.apache.thrift.TException;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import org.apache.hive.com.google.guava.cache.Cache;
-import org.apache.hive.com.google.guava.cache.CacheBuilder;
-import org.apache.hive.com.google.guava.cache.RemovalListener;
-import org.apache.hive.com.google.guava.cache.RemovalNotification;
-import org.apache.hive.com.google.guava.util.concurrent.ThreadFactoryBuilder;
+import org.apache.hive.com.google.common.cache.Cache;
+import org.apache.hive.com.google.common.cache.CacheBuilder;
+import org.apache.hive.com.google.common.cache.RemovalListener;
+import org.apache.hive.com.google.common.cache.RemovalNotification;
+import org.apache.hive.com.google.common.util.concurrent.ThreadFactoryBuilder;
 
 /**
  * A thread safe time expired cache for HiveMetaStoreClient
