+++ hcatalog/core/src/main/java/org/apache/hive/hcatalog/data/HCatRecordObjectInspectorFactory.java
@@ -22,8 +22,8 @@ import java.util.ArrayList;
 import java.util.List;
 import java.util.concurrent.TimeUnit;
 
-import org.apache.hive.com.google.guava.cache.Cache;
-import org.apache.hive.com.google.guava.cache.CacheBuilder;
+import org.apache.hive.com.google.common.cache.Cache;
+import org.apache.hive.com.google.common.cache.CacheBuilder;
 import org.apache.hadoop.hive.serde2.SerDeException;
 import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspector;
 import org.apache.hadoop.hive.serde2.objectinspector.ObjectInspectorFactory;
